Refactor code
Earlier, the app worked in a way that in addition to sending REST request, the app also kept
modifying a note array that was rendered. Instead of constantly fetching the notes from the
backend, the array was also modified.

This update changes this in a way that now the notes are fetched each time we go to the list.
Operations are done only to the backend and not the rendered list anymore.